+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 
 # Todo
+- sqlmath - bugfix - Fix 762-character-limit SQL-string-bug in function db_exec().
 - sqlmath - Re-introduce sliding-discrete-fourier-transform to improve initial guess in sql-function WIN_SINEFIT2().
 - sqlite - Add similar error-handling from builtin-sql-function PERCENTILE() into custom-sql-function QUANTILE().
 - none
+
+# v2026.3.1-beta
+- jslint-ci - Update shell-function shLintPython().
+- python - Add file test_lgbm_sqlite.py.
+- python - Auto-detect-and-load lib_lightgbm.dll.
+- sqlite - Add sqlite-extension csv.c, tsv.c.
 
 # v2026.2.28
 - python-ci - Fix ruff lint-errors.